Abstract ReporterBenchmark tests
