+++ core/build.gradle
@@ -30,8 +30,8 @@ sourceSets {
 }
 
 dependencies {
-    jmhImplementation 'org.openjdk.jmh:jmh-core:1.23'
-    jmhImplementation 'org.openjdk.jmh:jmh-generator-annprocess:1.23'
+    compile('org.openjdk.jmh:jmh-core:1.27')
+    compile('org.openjdk.jmh:jmh-generator-annprocess:1.27')
 }
 
 task jmh(type: JavaExec, dependsOn: jmhClasses) {
