--- core/src/jmh/java/com/uber/m3/tally/ReporterTemplateBenchmark.java
@@ -1,4 +1,0 @@
-package com.uber.m3.tally;
-
-public class ReporterTemplateBenchmark {
-}
