+++ prometheus/src/jmh/java/com/uber/m3/tally/prometheus/PrometheusReporterBenchmark.java
@@ -20,138 +20,25 @@
 
 package com.uber.m3.tally.prometheus;
 
-import com.uber.m3.tally.DurationBuckets;
+import com.uber.m3.jmh.AbstractReporterBenchmark;
 import com.uber.m3.tally.StatsReporter;
-import com.uber.m3.tally.ValueBuckets;
-import com.uber.m3.util.Duration;
-import com.uber.m3.util.ImmutableMap;
 import io.prometheus.client.CollectorRegistry;
-import org.openjdk.jmh.annotations.Benchmark;
 import org.openjdk.jmh.annotations.BenchmarkMode;
 import org.openjdk.jmh.annotations.Fork;
 import org.openjdk.jmh.annotations.Mode;
 import org.openjdk.jmh.annotations.OutputTimeUnit;
-import org.openjdk.jmh.annotations.Scope;
-import org.openjdk.jmh.annotations.Setup;
-import org.openjdk.jmh.annotations.State;
-import org.openjdk.jmh.annotations.TearDown;
 
 import java.util.concurrent.TimeUnit;
 
 @BenchmarkMode(Mode.Throughput)
 @OutputTimeUnit(TimeUnit.MILLISECONDS)
 @Fork(value = 2, jvmArgsAppend = {"-server", "-XX:+UseG1GC"})
-public class PrometheusReporterBenchmark {
+public class PrometheusReporterBenchmark extends AbstractReporterBenchmark {
 
-    private static final ImmutableMap<String, String> DEFAULT_TAGS = new ImmutableMap.Builder<String, String>(5)
-            .put("tag1", "test1")
-            .put("tag2", "test2")
-            .put("tag3", "test3")
-            .put("tag4", "test4")
-            .put("env", "test")
-            .build();
-
-    private static final DurationBuckets DURATION_EXPONENTIAL_BUCKETS = DurationBuckets.linear(
-            Duration.ofMillis(1), Duration.ofMillis(10), 128
-    );
-
-    private static final ValueBuckets VALUE_EXPONENTIAL_BUCKETS = ValueBuckets.linear(
-            0.1, 100d, 128
-    );
-
-    private static final String COUNTER_NAME = "counter";
-    private static final String GAUGE_NAME = "gauge";
-    private static final String TIMER_NAME = "timer";
-    private static final String HISTOGRAM_DURATION_NAME = "histogram_duration";
-    private static final String HISTOGRAM_VALUE_NAME = "histogram_value";
-
-    @Benchmark
-    public void reportCounterBenchmark(BenchmarkState state, IncrementalValueState value) {
-        state.reporter.reportCounter(COUNTER_NAME, DEFAULT_TAGS, value.incrementAndGet());
-    }
-
-    @Benchmark
-    public void reportGaugeBenchmark(BenchmarkState state, IncrementalValueState value) {
-        state.reporter.reportCounter(COUNTER_NAME, DEFAULT_TAGS, value.incrementAndGet());
-    }
-
-    @Benchmark
-    public void reportTimerBenchmark(BenchmarkState state, IncrementalValueState value) {
-        state.reporter.reportTimer(TIMER_NAME, DEFAULT_TAGS, Duration.ofSeconds(value.incrementAndGet()));
-    }
-
-    @Benchmark
-    public void reportHistogramDurationSamplesBenchmark(BenchmarkState state, IncrementalValueState value) {
-        state.reporter.reportHistogramDurationSamples(
-                HISTOGRAM_DURATION_NAME,
-                DEFAULT_TAGS,
-                DURATION_EXPONENTIAL_BUCKETS,
-                DURATION_EXPONENTIAL_BUCKETS.getDurationLowerBoundFor(10),
-                DURATION_EXPONENTIAL_BUCKETS.getDurationLowerBoundFor(10),
-                value.incrementAndGet()
-        );
-    }
-
-    @Benchmark
-    public void reportHistogramValueSamplesBenchmark(BenchmarkState state, IncrementalValueState value) {
-        state.reporter.reportHistogramValueSamples(
-                HISTOGRAM_VALUE_NAME,
-                DEFAULT_TAGS,
-                VALUE_EXPONENTIAL_BUCKETS,
-                5.0,
-                5.0,
-                value.incrementAndGet()
-        );
-    }
-
-    @State(Scope.Benchmark)
-    public static class BenchmarkState {
-
-        private StatsReporter reporter;
-
-        @Setup
-        public void setup() {
-            this.reporter = PrometheusReporter.builder()
-                    .registry(CollectorRegistry.defaultRegistry)
-                    .build();
-
-            reporter.reportCounter(COUNTER_NAME, DEFAULT_TAGS, 1);
-            reporter.reportGauge(GAUGE_NAME, DEFAULT_TAGS, 1);
-            reporter.reportTimer(TIMER_NAME, DEFAULT_TAGS, Duration.ofSeconds(1));
-            reporter.reportHistogramValueSamples(
-                    HISTOGRAM_VALUE_NAME,
-                    DEFAULT_TAGS,
-                    VALUE_EXPONENTIAL_BUCKETS,
-                    0.1,
-                    0.1,
-                    1
-            );
-            reporter.reportHistogramDurationSamples(
-                    HISTOGRAM_DURATION_NAME,
-                    DEFAULT_TAGS,
-                    DURATION_EXPONENTIAL_BUCKETS,
-                    Duration.ofSeconds(1),
-                    Duration.ofSeconds(1),
-                    1
-            );
-        }
-
-        @TearDown
-        public void teardown() {
-            reporter.close();
-        }
-
-    }
-
-    @State(Scope.Thread)
-    public static class IncrementalValueState {
-        private long x;
-
-        long incrementAndGet() {
-            if (x == Long.MAX_VALUE) {
-                x = 0;
-            }
-            return x++;
-        }
+    @Override
+    public StatsReporter initReporter() {
+        return PrometheusReporter.builder()
+                .registry(CollectorRegistry.defaultRegistry)
+                .build();
     }
 }
